refactored user auth component
