+++ packages/auth/lib/index.tsx
@@ -1,5 +1,8 @@
-export { AuthWrapper, AuthRoute, SignInCallback } from '../src/Auth';
-export { UserMenu, LoginButton, UserSubmissions, UserSettings } from '../src/User';
+export { AuthWrapper, AuthRoute, SignInCallback } from '../src/auth/Auth';
+export { UserMenu } from '../src/user/UserMenu';
+export { UserSubmissions } from '../src/user/UserSubmissions';
+export { UserSettings } from '../src/user/UserSettings';
+export * from '../src/user/Buttons';
 export { default as i18n } from '../src/languages/i18n';
 export { fetchUserProfile } from '../src/redux/store';
 export type { Target, AuthProperty } from '../src/types';
