+++ packages/auth/package.json
@@ -11,6 +11,7 @@
     "@reduxjs/toolkit": "^1.9.5",
     "i18next": "^23.4.1",
     "i18next-resources-to-backend": "^1.1.4",
+    "moment": "^2.29.4",
     "notistack": "^3.0.1",
     "oidc-client-ts": "^2.2.4",
     "react": "^18.2.0",
