+++ packages/auth/src/languages/locales/en/auth.json
@@ -1,0 +1,6 @@
+{
+	"signingIn": "Signing you in...",
+	"signingOut": "Signing you out...",
+	"isLoading": "Loading...",
+	"signinError": "Error signing you in!"
+}
