+++ packages/auth/src/languages/locales/nl/auth.json
@@ -1,0 +1,6 @@
+{
+	"signingIn": "Bezig met inloggen...",
+	"signingOut": "Bezig met uitloggen...",
+	"isLoading": "Momentje...",
+	"signinError": "Fout bij inloggen!"
+}
